Extract useTokenBalances to its own file
